Update to 1.18.2
Works in prod for both Fabric and Forge.
Does not work in dev for Forge and I'm not sure why.
I suspect that Fabric did some mismatch on intermediary or something, because this required zero changes code-wise, only buildscript -- nonetheless, this PR allows FC to be used on Fabric 1.18.2
I didn't test if Forge in prod was problematic before this, but I wouldn't think so.