Remove "warning" about mojmap
The warning is from a time when Mojmap was still new, and a significant
fraction of modders avoided it due to legal concerns. This has changed
in the last few years, and given that most mods are in Mojmap these days
it is no longer something that needs a warning.